refactor: step2 피드백 적용 - jgrapht 라이브러리 의존성 제거 및 일급컬렉션 생성
