+++ src/test/java/nextstep/subway/path/domain/PathFinderTest.java
@@ -1,6 +1,7 @@
 package nextstep.subway.path.domain;
 
 import nextstep.subway.line.domain.Line;
+import nextstep.subway.line.domain.Lines;
 import nextstep.subway.station.domain.Station;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.DefaultWeightedEdge;
@@ -10,6 +11,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -36,14 +38,10 @@ class PathFinderTest {
         Line 이호선 = new Line("이호선", "green", 교대역, 강남역, 10);
         Line 삼호선 = new Line("삼호선", "orange", 교대역, 양재역, 5);
         삼호선.addSection(교대역, 남부터미널역, 2);
-
-        List<Line> lines = new ArrayList<>();
-        lines.add(신분당선);
-        lines.add(이호선);
-        lines.add(삼호선);
+        Lines lines  = new Lines(Arrays.asList(신분당선, 이호선, 삼호선));
 
         // when
-        PathFinder pathFinder = new PathFinder(lines);
+        PathFinder pathFinder = new PathFinder(lines.createPath());
 
         // then
         assertThat(pathFinder.findPath(교대역, 양재역).getStations()).containsExactly(교대역, 남부터미널역, 양재역);
